Todo delete function complete

+++ server/app/api/todo.py
@@ -35,6 +35,7 @@ def createTodo():
 
         if 'title' not in body.keys():
             raise Exception('Required properties not specified')
+
         title = body['title']
 
         if 'description' not in body.keys():
@@ -54,6 +55,7 @@ def createTodo():
         return api_response(False, 'Todo created successfully', todo.serialize(fields))
     except Exception as error:
         return api_response(True, 'Failed to create todo', str(error))
+
 
 @todo_bp.route('/updateTodo/<int:id>', methods=['PATCH'])
 @jwt_required(optional=True)
@@ -90,15 +92,35 @@ def updateTodo(id: int):
     except Exception as error:
         return api_response(True, 'Failed to update todo', str(error))
 
-@todo_bp.route('/deleteTodo/<int:id>', methods=['DELETE'])
-def deleteTodo(id: int):
+
+@todo_bp.route('/deleteTodos', methods=['DELETE'])
+@jwt_required(optional=True)
+def deleteTodos():
     try:
-        todo = Todo.query.filter_by(id=id).first()
-        if not todo:
-            raise Exception('Object does not exist')
-            
-        db.session.delete(todo)
+        claims = get_jwt()
+        body = request.get_json()
+        
+        if 'ids' not in body.keys():
+            raise Exception('Todo ids not provided')
+
+        user_id = claims['id'] if 'id' in claims.keys() else None
+
+        todo_ids = body['ids']
+        deleted_ids = []
+        for todo_id in todo_ids:
+
+            todo = Todo.query.get(todo_id)
+            if not todo:
+                raise Exception('Object does not exist')
+
+            if todo.uid and todo.uid != user_id:
+                continue
+                # raise Exception('Insufficient permissions')
+
+            deleted_ids.append(todo.id)
+            db.session.delete(todo)
+
         db.session.commit()
-        return api_response(False, 'Successfully deleted todo')
+        return api_response(False, 'Successfully deleted todo', deleted_ids)
     except Exception as error:
         return api_response(True, 'Failed to delete todo', str(error))
